% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{foobar}
\alias{foobar}
\alias{foobar.numeric}
\alias{foobar.character}
\title{Foo bar generic}
\usage{
foobar(x)

\method{foobar}{numeric}(x)

\method{foobar}{character}(x)
}
\arguments{
\item{x}{Object to foo.}
}
\description{
Foo bar generic
}
\section{Methods (by class)}{
\itemize{
\item \code{foobar(numeric)}: Difference between the mean and the median

\item \code{foobar(character)}: First and last values pasted together in a string.

}}
