% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print hello world with a name}
\usage{
hello(name)
}
\arguments{
\item{name, }{A name to welcome.}
}
\description{
\code{hello} is a demonstration of R package function and this line should be
created in the "Description" section
}
\details{
This a the "Details" section for function \code{hello}
}
\examples{
hello(name='Xujia Liu')
#[1] "Hello, world! Xujia Liu"
}
\seealso{
\href{https://xujialiu.github.io/}{Xujia' s website} is a url link redirect to a website

\code{\link[=hello0]{hello0()}} is a link redirect to another function of this package

\code{\link[ggplot2:aes]{ggplot2::aes()}} is a link to another package function
}
