% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello0}
\alias{hello0}
\title{Print hello world}
\usage{
hello0()
}
\description{
\code{hello0} is a demonstration of R package function and this line should be
created in the "Description" section.
}
\details{
This a the "Details" section for function \code{hello0}
}
\seealso{
\href{https://xujialiu.github.io/}{Xujia' s website}is a url link redirect to a website
}
