% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrp-package.R
\docType{package}
\name{lrp-package}
\alias{lrp}
\alias{lrp-package}
\title{lrp: A R package for Xujia Liu to learn create a R package}
\description{
This is a useless package for anyone but Xujia Liu, the author and creater to learning create a R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://xujialiu.github.io/}
  \item \url{https://xujialiu.gitee.io/}
  \item \url{https://github.com/xujialiuphd/lrp}
  \item Report bugs at \url{https://github.com/xujialiuphd/lrp/issues}
}

}
\keyword{internal}
