% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrp.R
\name{lrp}
\alias{lrp}
\title{Représentation graphique de la relevance
Représentation d'une couche caché
Les fléches correspondent à la distribution de 50\% de la relevance}
\usage{
lrp(test, model, individu)
}
\arguments{
\item{test}{une matrice de données test}

\item{model}{modele keras utilisé}

\item{individu}{le numéro de l'individu qu'on veut tester}
}
\value{
un graphique de couche cachées
}
\description{
Représentation graphique de la relevance
Représentation d'une couche caché
Les fléches correspondent à la distribution de 50\% de la relevance
}
